requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
tqdm
yacs
